chore(admin): isolate ai apis to ee folder (#24834)
