+++ packages/core/admin/admin/src/index.ts
@@ -65,8 +65,6 @@ export { useMediaQuery, useIsDesktop, useIsTablet, useIsMobile } from './hooks/u
 export { useDeviceType } from './hooks/useDeviceType';
 export { useAdminUsers } from './services/users';
 export { useGetCountDocumentsQuery } from './services/homepage';
-/** @internal */
-export { useAIAvailability } from './hooks/useAIAvailability';
 
 /**
  * Types
