+++ packages/core/admin/ee/server/src/ai/controllers/__tests__/ai.test.ts
@@ -4,18 +4,8 @@ import createContext from '../../../../../../../../../tests/helpers/create-conte
 import aiController from '../ai';
 
 describe('AI Controller', () => {
-  const ORIGINAL_ENV = process.env;
-
   beforeEach(() => {
     jest.clearAllMocks();
-    process.env = { ...ORIGINAL_ENV }; // fresh copy for each test
-
-    // Reset global fetch
-    delete (global as any).fetch;
-  });
-
-  afterAll(() => {
-    process.env = ORIGINAL_ENV; // fully restore after suite
   });
 
   describe('getAiToken', () => {
@@ -32,52 +22,24 @@ describe('AI Controller', () => {
         {
           state: { user },
           unauthorized: jest.fn(),
-          badRequest: jest.fn(),
-          forbidden: jest.fn(),
-          notFound: jest.fn(),
           internalServerError: jest.fn(),
-          requestTimeout: jest.fn(),
           ...overrides,
         }
       );
     };
 
-    const createMockStrapi = (config = {}) => {
+    const createMockStrapi = (aiContainer = {}) => {
       return {
-        ee: { isEE: true },
-        dirs: { app: { root: '/app' } },
-        config: {
-          get: jest.fn((key, defaultValue) => {
-            if (key === 'uuid') return 'test-project-id';
-            return defaultValue;
-          }),
-        },
-        log: {
-          info: jest.fn(),
-          error: jest.fn(),
-          http: jest.fn(),
-        },
-        ...config,
+        get: jest.fn((service) => {
+          if (service === 'ai') {
+            return {
+              getAiToken: jest.fn(),
+              ...aiContainer,
+            };
+          }
+          return {};
+        }),
       };
-    };
-
-    // Helper to setup standard valid environment
-    const setupValidEnvironment = () => {
-      process.env.STRAPI_LICENSE = 'test-license';
-      process.env.STRAPI_AI_URL = 'http://ai-server.com';
-    };
-
-    // Helper to create successful AI server response
-    const createSuccessfulFetch = (responseData = {}) => {
-      const defaultResponse = {
-        jwt: 'test-jwt-token',
-        expiresAt: '2025-01-01T12:00:00Z',
-      };
-
-      return jest.fn().mockResolvedValueOnce({
-        ok: true,
-        json: jest.fn().mockResolvedValueOnce({ ...defaultResponse, ...responseData }),
-      });
     };
 
     test('Should return unauthorized when user is not authenticated', async () => {
@@ -89,82 +51,53 @@ describe('AI Controller', () => {
       expect(ctx.unauthorized).toHaveBeenCalledWith('Authentication required');
     });
 
-    test('Should return internal server error when no EE license is found', async () => {
+    test('Should return internal server error when AI container throws error', async () => {
       const ctx = createMockContext();
-      const mockStrapi = createMockStrapi() as any;
-      global.strapi = mockStrapi;
-
-      // Ensure no license is available
-      delete process.env.STRAPI_LICENSE;
+      const mockAiContainer = {
+        getAiToken: jest.fn().mockRejectedValue(new Error('Container error')),
+      };
+      global.strapi = createMockStrapi(mockAiContainer) as any;
 
       await aiController.getAiToken(ctx as any);
 
       expect(ctx.internalServerError).toHaveBeenCalledWith(
         'AI token request failed. Check server logs for details.'
       );
-
-      // Check that the specific error was logged
-      expect(mockStrapi.log.error).toHaveBeenCalledWith(
-        'AI token request failed: No EE license found. Please ensure STRAPI_LICENSE environment variable is set or license.txt file exists.'
-      );
+      expect(mockAiContainer.getAiToken).toHaveBeenCalled();
     });
 
-    test('Should return internal server error when project ID is not configured', async () => {
+    test('Should successfully return AI token when container returns token', async () => {
       const ctx = createMockContext();
-      const mockStrapi = createMockStrapi({
-        config: {
-          get: jest.fn((key) => (key === 'uuid' ? null : 'default-value')),
-        },
-      }) as any;
-      global.strapi = mockStrapi;
-      setupValidEnvironment();
-
-      await aiController.getAiToken(ctx as any);
-
-      expect(ctx.internalServerError).toHaveBeenCalledWith(
-        'AI token request failed. Check server logs for details.'
-      );
-
-      // Check that the specific error was logged
-      expect(mockStrapi.log.error).toHaveBeenCalledWith(
-        'AI token request failed: Project ID not configured'
-      );
-    });
-
-    test('Should successfully return AI token when all conditions are met', async () => {
-      const ctx = createMockContext();
-      const mockStrapi = createMockStrapi() as any;
-      global.strapi = mockStrapi;
-      setupValidEnvironment();
-
-      global.fetch = createSuccessfulFetch();
-
-      await aiController.getAiToken(ctx as any);
-
-      expect(global.fetch).toHaveBeenCalledWith(
-        'http://ai-server.com/auth/getAiJWT',
-        expect.objectContaining({
-          method: 'POST',
-          headers: expect.objectContaining({
-            'Content-Type': 'application/json',
-          }),
-          body: expect.stringContaining('test-license'),
-        })
-      );
-
-      expect(ctx.body).toEqual({
-        data: {
-          token: 'test-jwt-token',
-          expiresAt: '2025-01-01T12:00:00Z',
-        },
-      });
-
-      // Check that the appropriate log messages were called
-      expect(mockStrapi.log.http).toHaveBeenCalledWith('Contacting AI Server for token generation');
-      expect(mockStrapi.log.info).toHaveBeenCalledWith('AI token generated successfully', {
-        userId: 1,
+      const mockTokenData = {
+        token: 'test-jwt-token',
         expiresAt: '2025-01-01T12:00:00Z',
+      };
+      const mockAiContainer = {
+        getAiToken: jest.fn().mockResolvedValue(mockTokenData),
+      };
+      global.strapi = createMockStrapi(mockAiContainer) as any;
+
+      await aiController.getAiToken(ctx as any);
+
+      expect(mockAiContainer.getAiToken).toHaveBeenCalled();
+      expect(ctx.body).toEqual({
+        data: mockTokenData,
       });
+    });
+
+    test('Should delegate to AI container and return formatted response', async () => {
+      const ctx = createMockContext();
+      const mockAiContainer = {
+        getAiToken: jest
+          .fn()
+          .mockResolvedValue({ token: 'test-token', expiresAt: '2025-01-01T12:00:00Z' }),
+      };
+      global.strapi = createMockStrapi(mockAiContainer) as any;
+
+      await aiController.getAiToken(ctx as any);
+
+      expect(global.strapi.get).toHaveBeenCalledWith('ai');
+      expect(mockAiContainer.getAiToken).toHaveBeenCalled();
     });
   });
 });
