+++ packages/core/admin/ee/server/src/ai/routes/ai.ts
@@ -1,7 +1,6 @@
 export default {
   type: 'admin',
   routes: [
-    // ai data usage
     {
       method: 'GET',
       path: '/ai-usage',
@@ -10,7 +9,6 @@ export default {
         policies: ['admin::isAuthenticatedAdmin'],
       },
     },
-    // get ai token
     {
       method: 'GET',
       path: '/ai-token',
@@ -19,7 +17,6 @@ export default {
         policies: ['admin::isAuthenticatedAdmin'],
       },
     },
-    // get ai feature settings
     {
       method: 'GET',
       path: '/ai-feature-config',
