+++ packages/core/admin/server/src/routes/users.ts
@@ -24,14 +24,6 @@ export default [
     },
   },
   {
-    method: 'GET',
-    path: '/users/me/ai-token',
-    handler: 'authenticated-user.getAiToken',
-    config: {
-      policies: ['admin::isAuthenticatedAdmin'],
-    },
-  },
-  {
     method: 'POST',
     path: '/users',
     handler: 'user.create',
