+++ packages/core/upload/server/src/services/ai-metadata.ts
@@ -65,7 +65,7 @@ const createAIMetadataService = ({ strapi }: { strapi: Core.Strapi }) => {
 
       let token: string;
       try {
-        const tokenData = await strapi.service('admin::user').getAiToken();
+        const tokenData = await strapi.get('ai').getAiToken();
         token = tokenData.token;
       } catch (error) {
         throw new Error('Failed to retrieve AI token', {
