+++ packages/plugins/i18n/server/src/services/ai-localizations.ts
@@ -152,7 +152,7 @@ const createAILocalizationsService = ({ strapi }: { strapi: Core.Strapi }) => {
 
       let token: string;
       try {
-        const tokenData = await strapi.service('admin::user').getAiToken();
+        const tokenData = await strapi.get('ai').getAiToken();
         token = tokenData.token;
       } catch (error) {
         await aiLocalizationJobsService.upsertJobForDocument({
